requirements.txt
streamlit
instructor
openai
tiktoken
assemblyai
python-dotenv